# [Windows] Loading speedup (pcapdnet cache) + cleanup