… scapy/layers/inet6.py
… scapy/route.py
… scapy/utils.py
… test/fields.uts
… test/mock_windows.uts
… test/regression.uts