This PR speeds up windows booting (by a lot on windows 7):
- pcapdnet cache
- remove useless function
- do not read routes twice in `get_working_if`
- fixes route_add_loopback
- re-uses the same powershell instance while booting by adding a Powershell Manager
- fix long IP lists (especially IPv6 possible IPs): do not crop
- minor bug fixes (python3)

Results:
```
>>> a = time.time()
...:for i in range(1,10):
...:    read_routes[6]()
...:print(time.time()-a)
```
**on Windows 10:**

|  | Normal | After `POWERSHELL_PROCESS.__init__()` (=during of after booting) |
| --|---|---|
| 10 x read_routes | 8.33999991417s | 0.571000051498s |
| 10 x read_routes6 | 9.13199996948s | 0.917000055313s |
| 10 x get_windows_if_list | 12.8439998627s | 1.46499991417s |

POWERSHELL_PROCESS auto-turns off when scapy has finished booting (this could be changed, but do we want to leave a background task powershell open ?)