@rgom 2017-10-30
It hangs in GUI (PyDev). Last lines executed (python -m trace --trace):

```
 --- modulename: subprocess, funcname: _close_in_parent
subprocess.py(628):                 fd.Close()
subprocess.py(629):                 to_close.remove(fd)
subprocess.py(658):                 if errwrite is not None:
subprocess.py(659):                     _close_in_parent(errwrite)
 --- modulename: subprocess, funcname: _close_in_parent
subprocess.py(628):                 fd.Close()
subprocess.py(629):                 to_close.remove(fd)
subprocess.py(662):             self._child_created = True
subprocess.py(663):             self._handle = hp
subprocess.py(664):             self.pid = pid
subprocess.py(665):             ht.Close()
subprocess.py(406):         if mswindows:
subprocess.py(407):             if p2cwrite is not None:
subprocess.py(409):             if c2pread is not None:
subprocess.py(410):                 c2pread = msvcrt.open_osfhandle(c2pread.Detach(), 0)
subprocess.py(411):             if errread is not None:
subprocess.py(414):         if p2cwrite is not None:
subprocess.py(416):         if c2pread is not None:
subprocess.py(417):             if universal_newlines:
subprocess.py(418):                 self.stdout = os.fdopen(c2pread, 'rU', bufsize)
subprocess.py(421):         if errread is not None:
__init__.py(78):     l=[]
__init__.py(79):     for line in ps.stdout:
```

I'm not sure, but that does seem to be a regression from master (or at least there it was harder to observe).

Robert