@gpotter2 2017-10-30
@rgom Hi,
Please note that this isn't `win-speedup-2`, as it does not provides the powershell fix. It also passes the tests on AppVeyor, which means it was able to boot and load correctly.

Regards