@rgom 2017-10-30
I'm not sure if I understand your recent input (which I appreciate), so let me elaborate:
1. Newer scapy versions can hang on loading when used in a script called from IDE.
2. With this version of scapy I'm suffering from that as well (so far 3/3 reproducibility)
3. At the same time I'm able to import scapy (the same way) from console

So while AppVeyor checks are essential, they don't cover all use cases. And this particular use case can be yes-no to a library.

I think (guess) that communicate would help with that, but it probably won't fit other process model (communicate would like to wait for other process termination).

I can elaborate further, if needed.

Regards,
Robert