@gpotter2 2017-10-31
@rgom Thanks for your report.

I've done some researchs, and it looks like passing parameters `-Command -` to powershell should fix the "hanging" bug.

i've updated this PR with the full powershell manager, which now *should* be working properly...

Regards,
Gabriel