@rgom 2017-10-31
I'm impressed.
This change works for me and reduces loading time from 10-12 seconds to 4-5 seconds. That's really a gain.

As for the persistence of the process. Current way is fine for me, previous version with permanent powershell locked some files, so upgrade was problematic.
It could be changed to use atexit, if desired.

Regards,
Robert