@gpotter2 2017-11-01
@rgom Thanks for your comments !
I've updated the code