@gpotter2 2017-11-01
@p-l- Ready to be reviewed