@gpotter2 2017-11-02
The last commit reverses your fix, as appveyor fails on it.

Seems that's it breaks IPv6 routes loading... will investigate