@p-l- 2017-11-02
@gpotter2 if you need to cache some data for Windows, I think you should either:
  * try to prevent the data from being computed several times in a first place
  * or at least restrict the changes (if they cannot be avoided) to the Windows specific code.

Or is there also a problem with Unix Pcap/Dnet code?