@gpotter2 2017-11-02
@p-l- Alright, i've added a windows specific cache.

@rgom I've also added POWERSHELL_MANAGER calls to the windows 7 metric function, to speed up booting even more.