@rgom 2017-11-02
FYI - the recent version slowed down significantly.
Previously it was 4.5 s, now it's 7.5 s, with most of it spent in event.wait().
(Similar effect was in my experiments. I haven't found a root cause, but I believe there must be some threads starvation or inter-locking issue).
I suspect pcap_service_something, which hasn't been converted and launches a new process always.

Could you not amend commits and just add commits? That way we could refer to particular versions and whole github machine would support just fine.

And again - thanks for working on that - appreciated greatly.

Robert