@rgom 2017-11-02
I should make that clear - when I was talking about time, I was talking about time when profiling.
When running just normally the time is normally about 50% lower.
But that means if I was able to reach 3.8 s in profiling, I should be able to reach <2 when non profiling, right?

And a simple debugging in non-profiling session, showing that there's more room for improvement (times are in seconds):

```
>>> from scapy import all as scapy
powershell process running 1 ['Get-WmiObject', 'Win32_NetworkAdapter', '|', 'select Name, InterfaceIndex, InterfaceDescription, GUID, MacAddress, NetConnectionID', '|', 'fl'] 3.77556102005e-07
powershell process running 1 ['Get-WmiObject', 'Win32_NetworkAdapterConfiguration', '|', 'select Description, IPAddress', '|', 'fl'] 1.90437410071
powershell process running 1 ['Get-WmiObject', 'Win32_NetworkAdapterConfiguration', '|', 'select Description, IPAddress', '|', 'fl'] 1.98737905462
powershell process running 1 ['Get-WmiObject', 'Win32_NetworkAdapterConfiguration', '|', 'select Description, IPAddress', '|', 'fl'] 2.07036286539
powershell process running 1 ['Get-WmiObject', 'Win32_NetworkAdapterConfiguration', '|', 'select Description, IPAddress', '|', 'fl'] 2.16445324402
powershell process running 1 ['Get-WmiObject', 'Win32_NetworkAdapterConfiguration', '|', 'select Description, IPAddress', '|', 'fl'] 2.24914813904
powershell process running 1 ['Get-WmiObject', 'Win32_NetworkAdapterConfiguration', '|', 'select Description, IPAddress', '|', 'fl'] 2.32737096737
powershell process running 1 ['Get-WmiObject', 'Win32_IP4RouteTable', '|', 'select Name, Mask, NextHop, InterfaceIndex, Metric1', '|', 'fl'] 2.40888306446
powershell process running 3 2.48356328388
powershell process running 2 netsh interface ipv6 show interfaces level=verbose 11 3.13995400612
powershell process running 2 netsh interface ipv6 show interfaces level=verbose 11 3.34311203638
powershell process running 2 netsh interface ipv6 show interfaces level=verbose 11 3.54641542573
3.74936844303
```

Regards,
Robert