@rgom 2017-11-02
Another thing - this pattern is repeated over and over:

```
    if POWERSHELL_PROCESS.running:
        stdout = b"".join(POWERSHELL_PROCESS.query(query_cmd))
    else:
        stdout, _ = sp.Popen([conf.prog.cmd, '/c', query_cmd], stdout=sp.PIPE).communicate()
```

I think it could be created as a method in Powershell manager - less duplicated code.

HTH,
Robert