@gpotter2 2017-11-06
@rgom Hi, sorry for delay.
The latest commit should speed up the loading time:
- it converts pcap_...
- remove duplicated code + splitting