@rgom 2017-11-06
Thank you.
If you need my feedback, please give me a day or two to provide it.
Other than that with every iteration the code looks cleaner and more robust.
And providing individual commits really helps.

Robert